requirements.txt
torch>=2.0
python-chess>=1.9
zstandard>=0.21
parameterized
